# internal/docs/swagger.yaml
definitions:
...
  github_com_umar5678_go-backend_internal_models.UserRole:
    enum:
    - rider
    - driver
    - admin
    - delivery_person
    - service_provider
    - handyman
    type: string
    x-enum-varnames:
    - RoleRider
    - RoleDriver
    - RoleAdmin
    - RoleDeliveryPerson
    - RoleServiceProvider
    - RoleHandyman
  github_com_umar5678_go-backend_internal_models.UserStatus:
    enum:
    - active
    - suspended
    - banned
    - pending_verification
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusSuspended
    - StatusBanned
    - StatusPendingVerification
  github_com_umar5678_go-backend_internal_models.WalletType:
    enum:
    - rider
    - driver
    - platform
    type: string
    x-enum-varnames:
    - WalletTypeRider
    - WalletTypeDriver
    - WalletTypePlatform
...
  github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      name:
        type: string
      phone:
        type: string
      profilePhotoUrl:
        type: string
      role:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
      status:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserStatus'
    type: object
...
  github_com_umar5678_go-backend_internal_modules_pricing_dto.FareEstimateRequest:
    properties:
      dropoffLat:
        maximum: 90
        minimum: -90
        type: number
      dropoffLon:
        maximum: 180
        minimum: -180
        type: number
      pickupLat:
        maximum: 90
        minimum: -90
        type: number
      pickupLon:
        maximum: 180
        minimum: -180
        type: number
      vehicleTypeId:
        type: string
    required:
    - dropoffLat
    - dropoffLon
    - pickupLat
    - pickupLon
    - vehicleTypeId
    type: object
  github_com_umar5678_go-backend_internal_modules_pricing_dto.FareEstimateResponse:
    properties:
      baseFare:
        type: number
      bookingFee:
        type: number
      currency:
        type: string
      distanceFare:
        type: number
      durationFare:
        type: number
      estimatedDistance:
        description: km
        type: number
      estimatedDuration:
        description: seconds
        type: integer
      subTotal:
        type: number
      surgeAmount:
        type: number
      surgeMultiplier:
        type: number
      totalFare:
        type: number
      vehicleTypeName:
        type: string
    type: object
...
  github_com_umar5678_go-backend_internal_modules_riders_dto.AddressInput:
    properties:
      address:
        maxLength: 500
        minLength: 5
        type: string
      lat:
        type: number
      lng:
        type: number
    required:
    - address
    - lat
    - lng
    type: object
  github_com_umar5678_go-backend_internal_modules_riders_dto.AddressResponse:
    properties:
      address:
        type: string
      lat:
        type: number
      lng:
        type: number
    type: object
  github_com_umar5678_go-backend_internal_modules_riders_dto.RiderProfileResponse:
    properties:
      createdAt:
        type: string
      homeAddress:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.AddressResponse'
      id:
        type: string
      preferredVehicleType:
        type: string
      rating:
        type: number
      totalRides:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      userId:
        type: string
      wallet:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.WalletResponse'
      workAddress:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.AddressResponse'
    type: object
  github_com_umar5678_go-backend_internal_modules_riders_dto.RiderStatsResponse:
    properties:
      memberSince:
        type: string
      rating:
        type: number
      totalRides:
        type: integer
      walletBalance:
        type: number
    type: object
  github_com_umar5678_go-backend_internal_modules_riders_dto.UpdateProfileRequest:
    properties:
      homeAddress:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.AddressInput'
      preferredVehicleType:
        type: string
      workAddress:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.AddressInput'
    type: object
  github_com_umar5678_go-backend_internal_modules_rides_dto.CancelRideRequest:
    properties:
      reason:
        maxLength: 500
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_rides_dto.CompleteRideRequest:
    properties:
      actualDistance:
        minimum: 0
        type: number
      actualDuration:
        minimum: 0
        type: integer
    required:
    - actualDistance
    - actualDuration
    type: object
  github_com_umar5678_go-backend_internal_modules_rides_dto.CreateRideRequest:
    properties:
      dropoffAddress:
        maxLength: 500
        type: string
      dropoffLat:
        maximum: 90
        minimum: -90
        type: number
      dropoffLon:
        maximum: 180
        minimum: -180
        type: number
      pickupAddress:
        maxLength: 500
        type: string
      pickupLat:
        maximum: 90
        minimum: -90
        type: number
      pickupLon:
        maximum: 180
        minimum: -180
        type: number
      riderNotes:
        maxLength: 500
        type: string
      vehicleTypeId:
        type: string
    required:
    - dropoffLat
    - dropoffLon
    - pickupLat
    - pickupLon
    - vehicleTypeId
    type: object
  github_com_umar5678_go-backend_internal_modules_rides_dto.RejectRideRequest:
    properties:
      reason:
        maxLength: 500
        type: string
      rideId:
        type: string
    required:
    - rideId
    type: object
  github_com_umar5678_go-backend_internal_modules_rides_dto.RideListResponse:
    properties:
      dropoffAddress:
        type: string
      estimatedFare:
        type: number
      id:
        type: string
      pickupAddress:
        type: string
      requestedAt:
        type: string
      status:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse:
    properties:
      acceptedAt:
        type: string
      actualDistance:
        type: number
      actualDuration:
        type: integer
      actualFare:
        type: number
      arrivedAt:
        type: string
      cancellationReason:
        type: string
      cancelledAt:
        type: string
      cancelledBy:
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      driver:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      driverId:
        type: string
      dropoffAddress:
        type: string
      dropoffLat:
        type: number
      dropoffLon:
        type: number
      estimatedDistance:
        type: number
      estimatedDuration:
        type: integer
      estimatedFare:
        type: number
      id:
        type: string
      pickupAddress:
        type: string
      pickupLat:
        type: number
      pickupLon:
        type: number
      requestedAt:
        type: string
      rider:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      riderId:
        type: string
      riderNotes:
        type: string
      startedAt:
        type: string
      status:
        type: string
      surgeMultiplier:
        type: number
      updatedAt:
        type: string
      vehicleType:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse'
      vehicleTypeId:
        type: string
    type: object
...
  github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse:
    properties:
      baseFare:
        type: number
      bookingFee:
        type: number
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      displayName:
        type: string
      iconUrl:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      perKmRate:
        type: number
      perMinuteRate:
        type: number
    type: object
...
  github_com_umar5678_go-backend_internal_modules_wallet_dto.WalletResponse:
    properties:
      availableBalance:
        type: number
      balance:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      heldBalance:
        type: number
      id:
        type: string
      isActive:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      userId:
        type: string
      walletType:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.WalletType'
    type: object
...
  github_com_umar5678_go-backend_internal_utils_response.ErrorDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Meta:
    properties:
      durationMs:
        type: number
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Response:
    properties:
      code:
        type: string
      data: {}
      errors:
        description: Use ErrorDetail
        items:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.ErrorDetail'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Meta'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready Go backend API
  title: supr booking server in go
  version: "1.0"
paths:
...
  /pricing/estimate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Fare estimate request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_pricing_dto.FareEstimateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_pricing_dto.FareEstimateResponse'
              type: object
      summary: Get fare estimate for a trip
      tags:
      - pricing
  /pricing/surge:
    get:
      parameters:
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: Get surge multiplier for a location
      tags:
      - pricing
...
  /riders/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.RiderProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get rider profile
      tags:
      - riders
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.RiderProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update rider profile
      tags:
      - riders
  /riders/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_riders_dto.RiderStatsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get rider statistics
      tags:
      - riders
  /rides:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: User role (rider or driver)
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideListResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List user's rides
      tags:
      - rides
    post:
      consumes:
      - application/json
      parameters:
      - description: Ride request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.CreateRideRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new ride request
      tags:
      - rides
  /rides/{id}:
    get:
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get ride details
      tags:
      - rides
  /rides/{id}/accept:
    post:
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Accept a ride request (Driver)
      tags:
      - rides
  /rides/{id}/arrived:
    post:
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Mark driver as arrived at pickup (Driver)
      tags:
      - rides
  /rides/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Cancellation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.CancelRideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Cancel a ride
      tags:
      - rides
  /rides/{id}/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Completion data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.CompleteRideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Complete the ride (Driver)
      tags:
      - rides
  /rides/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RejectRideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Reject a ride request (Driver)
      tags:
      - rides
  /rides/{id}/start:
    post:
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_rides_dto.RideResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Start the ride (Driver)
      tags:
      - rides
...
swagger: "2.0"